refactor(sponsor): get register page working again
