+++ src/components/register/register-member-payment.tsx
@@ -2,7 +2,6 @@
 import {
   Box,
   Card,
-  Divider,
   FormControl,
   FormControlLabel,
   Radio,
